Added picasso and images to recyclerView
Experimented with search functionality but have taken that out for the meanwhile. Went ahead and added images to the recyclerView. I did this by updating the firebase database with image sources from the same data source and using picasso to read them from the internet. This does mean the app will require internet access to display images.