Added timers to user plants
Have the basic functionality implemented. The timer is set used two calender views and an intent, and i'm happy to say it can be canceled before from datepicker.xml. These timers have no functionality yet but I will want to use this information and the announcements page to allow users to interact with the management side of their plants.