Added icon and more
Created a simple icon for the application. I am also working out some issues with the calendar views. They are not fixed yet but this commit shows some promise in resolving the problem. The problem is that the time getting stored in the local database that I want to use as a Timer is not storing the correct times that a user enters on the calendar.