Changed the way that a timer is added
Finally removed the calendar selection/datepicker.xml intent that I was trying to use to add the timers before. After debugging it and trying everything I could find to try I was unable to get the correct data from the intent. Therefore I have changed the method of adding a timer/schedule to the plant. Now when a user adds a plant there will be an EditText option where they can enter how many days they want between each watering. The timer will then be set assuming the plant was just watered. From that point on we will use the updateTimer method in the plantdatabasehelper in order to update the timer when the user clicks some button indicating they just watered it. If the user neglects to enter a number of days then it will be assumed they do not want to track this plant via timer and one will not be set. This information is used to add a calendar to the local plant detail view indicating when the next watering event will be. If the current time exceeds the scheduled time then the calendar will be red. Additionally, I cleaned up the files by removing unused/old imports and removed the viewModels as i have not used them in my solutions.